+++ internal/services/filesharer/internal/handler/incoming/load_test.go
@@ -70,6 +70,9 @@ func TestHandleIncomingLoadHTTP(t *testing.T) {
 	if err := incomingLoadRequestInvalidPersonal(handler); err != nil {
 		t.Fatal(err)
 	}
+	if err := incomingLoadRequestInvalidFileName(handler); err != nil {
+		t.Fatal(err)
+	}
 	if err := incomingLoadRequestGetSharingStorage(handler); err != nil {
 		t.Fatal(err)
 	}
@@ -97,7 +100,26 @@ func incomingLoadRequestGetSharingStorage(handler http.HandlerFunc) error {
 
 func incomingLoadRequestInvalidPersonal(handler http.HandlerFunc) error {
 	w := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodGet, "/?personal=qwerty", nil)
+	req := httptest.NewRequest(http.MethodGet, "/load?name=file.txt&chunk=0&personal=qwerty", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer func() { _ = res.Body.Close() }()
+
+	if res.StatusCode != http.StatusBadRequest {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func incomingLoadRequestInvalidFileName(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/?name=", nil)
 
 	handler(w, req)
 	res := w.Result()
@@ -135,7 +157,7 @@ func incomingLoadRequestBigChunk(handler http.HandlerFunc) error {
 
 func incomingLoadRequestNotFound(handler http.HandlerFunc) error {
 	w := httptest.NewRecorder()
-	req := httptest.NewRequest(http.MethodGet, "/load?name=norfound.txt&chunk=0", nil)
+	req := httptest.NewRequest(http.MethodGet, "/load?name=not_found.txt&chunk=0", nil)
 
 	handler(w, req)
 	res := w.Result()