+++ internal/services/filesharer/internal/handler/local_file_info.go
@@ -31,10 +31,10 @@ func HandleLocalFileInfoAPI(
 		}
 
 		queryParams := pR.URL.Query()
+		fileName := queryParams.Get("name")
 		aliasName := queryParams.Get("friend")
 
-		fileName := filepath.Base(queryParams.Get("name"))
-		if fileName == "" || fileName != queryParams.Get("name") {
+		if utils.FileNameIsInvalid(fileName) {
 			pLogger.PushWarn(logBuilder.WithMessage("got_invalid_name"))
 			_ = api.Response(pW, http.StatusBadRequest, "failed: got invalid name")
 			return