+++ internal/services/filesharer/internal/handler/remote_file_test.go
@@ -39,6 +39,9 @@ func TestHandleRemoteFileAPI(t *testing.T) {
 	if err := remoteFileRequestInvalidMethod(handlerX); err != nil {
 		t.Fatal(err)
 	}
+	if err := remoteFileRequestInvalidFileName(handlerX); err != nil {
+		t.Fatal(err)
+	}
 	if err := remoteFileRequestInvalidPersonal(handlerX); err != nil {
 		t.Fatal(err)
 	}
@@ -187,6 +190,21 @@ func remoteFileRequestInvalidMethod(handler http.HandlerFunc) error {
 	defer func() { _ = res.Body.Close() }()
 
 	if res.StatusCode != http.StatusMethodNotAllowed {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	return nil
+}
+
+func remoteFileRequestInvalidFileName(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/?name=", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer func() { _ = res.Body.Close() }()
+
+	if res.StatusCode != http.StatusBadRequest {
 		return errors.New("bad status code") // nolint: err113
 	}